retry rm volume on "in use" error
