+++ pipeline/backend/docker/docker.go
@@ -22,7 +22,7 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
-	"strings"
+	"time"
 
 	"github.com/containerd/errdefs"
 	"github.com/docker/go-connections/tlsconfig"
@@ -41,7 +41,11 @@ import (
 	"go.woodpecker-ci.org/woodpecker/v3/shared/utils"
 )
 
-var containerKillTimeout = 5 // seconds
+const (
+	containerKillTimeout = 5 // seconds
+	volumeRetryWait      = 1 // seconds
+	maxRetry             = 3
+)
 
 type docker struct {
 	client client.APIClient
@@ -329,7 +333,7 @@ func (e *docker) DestroyStep(ctx context.Context, step *backend_types.Step, task
 
 	// we first signal to the container to stop ...
 	if _, err := e.client.ContainerStop(ctx, containerName, client.ContainerStopOptions{
-		Timeout: &containerKillTimeout,
+		Timeout: toRef(containerKillTimeout),
 	}); err != nil && !isErrContainerNotFoundOrNotRunning(err) {
 		// we do not return error yet as we try to kill it first
 		stopErr = fmt.Errorf("could not stop container '%s': %w", step.Name, err)
@@ -367,11 +371,21 @@ func (e *docker) DestroyWorkflow(ctx context.Context, conf *backend_types.Config
 		log.Error().Err(err).Msgf("could not destroy all containers")
 	}
 
-	if _, err := e.client.VolumeRemove(ctx, conf.Volume, client.VolumeRemoveOptions{
-		Force: true,
-	}); err != nil {
+	var err error
+	for retryCount := 0; retryCount < maxRetry; retryCount++ {
+		_, err = e.client.VolumeRemove(ctx, conf.Volume, client.VolumeRemoveOptions{
+			Force: true,
+		})
+		if err == nil || !isErrVolumeInUse(err) {
+			// if it worked or if we have no in use error do not retry
+			break
+		}
+		time.Sleep(volumeRetryWait * time.Second)
+	}
+	if err != nil {
 		log.Error().Err(err).Msgf("could not remove volume '%s'", conf.Volume)
 	}
+
 	if _, err := e.client.NetworkRemove(ctx, conf.Network, client.NetworkRemoveOptions{}); err != nil {
 		log.Error().Err(err).Msgf("could not remove network '%s'", conf.Network)
 	}
@@ -382,19 +396,6 @@ var removeOpts = client.ContainerRemoveOptions{
 	RemoveVolumes: true,
 	RemoveLinks:   false,
 	Force:         false,
-}
-
-func isErrContainerNotFoundOrNotRunning(err error) bool {
-	// Error response from daemon: Cannot kill container: ...: No such container: ...
-	// Error response from daemon: Cannot kill container: ...: Container ... is not running"
-	// Error response from podman daemon: can only kill running containers. ... is in state exited
-	// Error response from daemon: removal of container ... is already in progress
-	// Error: No such container: ...
-	return err != nil &&
-		(strings.Contains(err.Error(), "No such container") ||
-			strings.Contains(err.Error(), "is not running") ||
-			strings.Contains(err.Error(), "can only kill running containers") ||
-			(strings.Contains(err.Error(), "removal of container") && strings.Contains(err.Error(), "is already in progress")))
 }
 
 // normalizeArchType converts the arch type reported by docker info into
